# api.yml
swagger: '2.0'
info:
  description: >-
    Kaori è un server che mette a disposizione API di servizio utili per gestire
    i servizi richiesti dal proprio client. Il server è utilizzabile anche da
    altri client ma non è pensato per essi per cui potrebbe non adempire in modo
    consono al servizio richiesto.
  version: 1.0.0
  title: Kaori API
  termsOfService: ''
  contact:
    email: brignoli.diego.studente@itispaleocapa.it
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'waifuai.dd-dns.de:8012'
basePath: /api
...
schemes:
  - https
paths:
  '/KaoriGui/{nomeFile}':
    get:
      summary: File per l'interfaccia grafica
      description: >-
        L'API KaoriGui funziona tramite metodo GET. È stata realizzata per
        servire i file dell'interfaccia grafica.
      operationId: kaoriGui
      parameters:
        - name: nomeFile
          in: path
          description: Nome del file
          required: true
          type: string
      responses:
        '200':
          description: Operazione completata con successo
        '404':
          description: File non trovato
...
  /auth/refresh:
    get:
      tags:
        - auth
      summary: Refresh dei token
      description: >-
        L'API di refresh funziona per metodo GET. Effettua il refresh
        dell'access token e del refresh token. Inoltre cotrolla nel database se
        ci sono refresh token dell'utente scaduti da eliminare.
      operationId: refresh
      parameters:
        - in: header
          name: Cookie
          type: string
          required: true
      responses:
        '200':
          description: Operazione completata con successo
          schema:
            type: object
            properties:
              AccessToken:
                type: string
                example: >-
                  eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJjb21wYW55IjoiQ29kZU9mVGhlS25pZ2h0IiwiZW1haWwiOiJ3YXRhc2hpd2F5dXJpZGFpc3VraUBnbWFpbC5jb20iLCJleHAiOjE2MjEwNzAxNjYsImlhdCI6MTYyMTA2OTI2NiwiaXNzIjoiS2FvcmlTdHJlYW0uY29tIiwicGVybWlzc2lvbiI6InVjdGEifQ.KkvK5XoBofs2HWSarlWdAdoaxjvnbKaoJmgwSn96QbY
              Expiration:
                type: integer
                format: int64
                example: 1621070166
          headers:
            Set-cookie:
              type: string
              description: contiene il refresh token in un cookie cifrato.
        '400':
          description: Bad request
          schema:
            $ref: '#/definitions/Response'
        '401':
          description: Token not valid
          schema:
            $ref: '#/definitions/Response'
        '500':
          description: Errore interno del server
          schema:
            $ref: '#/definitions/Response'
...
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: Authorization
definitions:
...
  Response:
    type: object
    properties:
      code:
        type: integer
        format: int32
      msg:
        type: string